chore: test-exception ci
