--- src/runtime/controller.test.ts
@@ -1,34 +1,0 @@
-import { validateHash, ValidationError } from "./controller";
-import { describe, it, expect, jest, beforeEach, afterEach } from "@jest/globals";
-
-describe("validateHash", () => {
-  let originalExit: (code?: number) => never;
-
-  beforeEach(() => {
-    originalExit = process.exit;
-    process.exit = jest.fn() as unknown as (code?: number) => never;
-  });
-
-  afterEach(() => {
-    process.exit = originalExit;
-  });
-  it("should throw ValidationError for invalid hash values", () => {
-    // Examples of invalid hashes
-    const invalidHashes = [
-      "", // Empty string
-      "12345", // Too short
-      "zxcvbnmasdfghjklqwertyuiop1234567890zxcvbnmasdfghjklqwertyuio", // Contains invalid character 'z'
-      "123456789012345678901234567890123456789012345678901234567890123", // 63 characters, one short
-    ];
-
-    invalidHashes.forEach(hash => {
-      expect(() => validateHash(hash)).toThrow(ValidationError);
-    });
-  });
-
-  it("should not throw ValidationError for valid SHA-256 hash", () => {
-    // Example of a valid SHA-256 hash
-    const validHash = "abc123def456abc123def456abc123def456abc123def456abc123def456abc1";
-    expect(() => validateHash(validHash)).not.toThrow();
-  });
-});
